Waypoints, pathing in and out of buildings, to specific tiles, and returning home at the end of day.
