+++ StarAI/StarAI/PathFindingCore/Utilities.cs
@@ -383,7 +383,7 @@ namespace StarAI.PathFindingCore
                     if (f == false)
                     {
                         ModCore.CoreMonitor.Log("FAILED TO PUT DOWN A GOAL????");
-                        ModCore.CoreMonitor.Log(v.thisLocation.ToString()+v.tileLocation.ToString());
+                        ModCore.CoreMonitor.Log(v.thisLocation.ToString()+pos.ToString());
                     }
                     // ModCore.CoreMonitor.Log("OK THIS IS THE RESULT F: " + f, LogLevel.Alert);
                     if (f == true)
@@ -436,52 +436,8 @@ namespace StarAI.PathFindingCore
                     //ModCore.CoreMonitor.Log("CAUGHT MY CULPERATE", LogLevel.Warn);
                 }
             }
-
-
-
-            /*
-            foreach (var nav in miniGoals) //change this??????
-            {
-                Utilities.clearExceptionListWithName("Child");
-                Utilities.clearExceptionListWithName("Navigation");
-                TileNode tempSource = new TileNode(1, Vector2.Zero, Path.Combine("Tiles", "GenericUncoloredTile.xnb"), Path.Combine("Tiles", "TileData.xnb"), StardustCore.IlluminateFramework.Colors.invertColor(StardustCore.IlluminateFramework.ColorsList.RosyBrown));
-               if(placement) tempSource.placementAction(Game1.player.currentLocation, Game1.player.getTileX() * Game1.tileSize, Game1.player.getTileY() * Game1.tileSize);
-                else tempSource.fakePlacementAction(Game1.player.currentLocation, Game1.player.getTileX(), Game1.player.getTileY());
-
-                Utilities.tileExceptionList.Add(new TileExceptionMetaData(tempSource, "Navigation"));
-                //StaardustCore.Utilities.masterAdditionList.Add(new StardustCore.DataNodes.PlacementNode(tempSource, Game1.currentLocation, Game1.player.getTileX() * Game1.tileSize, Game1.player.getTileY() * Game1.tileSize));
-
-
-                //have this take in a list of goals and see which goal it reaches first
-                List<TileNode> path = PathFindingCore.PathFindingLogic.pathFindToSingleGoalReturnPath(tempSource, nav, new List<TileNode>(), placement, utility);
-                if (path.Count == 0)
-                {
-                    ModCore.CoreMonitor.Log("NOPE, no path I guess.",LogLevel.Warn);
-                }
-                else
-                {
-                    ModCore.CoreMonitor.Log("There is a path", LogLevel.Alert);
-                    ModCore.CoreMonitor.Log("COST OF THE PATH IS: "+path.Count.ToString(), LogLevel.Alert);
-                }
-                if (path.Count != 0)
-                {
-                    //ModCore.CoreMonitor.Log("PATH WAS NOT NULL", LogLevel.Warn);
-                    paths.Add(path);
-                    foreach (var someTile in path)
-                    {
-                        if (someTile == nav) removalList.Add(someTile);
-                        StardustCore.ModCore.SerializationManager.trackedObjectList.Remove(someTile);
-                       if(placement) someTile.thisLocation.objects.Remove(someTile.tileLocation);
-                       
-                       //someTile.performRemoveAction(someTile.tileLocation, someTile.thisLocation);
-                        //StardustCore.Utilities.masterRemovalList.Add(someTile);
-                        //ModCore.CoreMonitor.Log("CAUGHT MY CULPERATE", LogLevel.Warn);
-                    }
-                }
-
-            }
-            */
-            Console.WriteLine("GOALS COUNT:" + miniGoals.Count);
+            
+           // Console.WriteLine("GOALS COUNT:" + miniGoals.Count);
             foreach (var q in removalList)
             {
                 StardustCore.ModCore.SerializationManager.trackedObjectList.Remove(q);
