+++ StarAI/StarAI/StarAI.csproj
@@ -70,19 +70,20 @@
     <Compile Include="ExecutionCore\TaskPrerequisites\StaminaPrerequisite.cs" />
     <Compile Include="ExecutionCore\TaskPrerequisites\ToolPrerequisite.cs" />
     <Compile Include="ModCore.cs" />
-    <Compile Include="PathFindingCore\ChestLogic.cs" />
-    <Compile Include="PathFindingCore\CropLogic\CropLogic.cs" />
-    <Compile Include="PathFindingCore\CropLogic\ShippingLogic.cs" />
-    <Compile Include="PathFindingCore\DebrisLogic\DebrisLogic.cs" />
-    <Compile Include="PathFindingCore\MapTransitionLogic\TransitionLogic.cs" />
-    <Compile Include="PathFindingCore\MapTransitionLogic\WarpGoal.cs" />
+    <Compile Include="TaskCore\ChestLogic.cs" />
+    <Compile Include="TaskCore\CropLogic\CropLogic.cs" />
+    <Compile Include="TaskCore\CropLogic\ShippingLogic.cs" />
+    <Compile Include="TaskCore\DebrisLogic\DebrisLogic.cs" />
+    <Compile Include="TaskCore\MapTransitionLogic\TransitionLogic.cs" />
+    <Compile Include="TaskCore\MapTransitionLogic\WarpGoal.cs" />
     <Compile Include="PathFindingCore\PathFindingLogic.cs" />
+    <Compile Include="TaskCore\MapTransitionLogic\WayPoints.cs" />
     <Compile Include="PathFindingCore\TileNodes\PlacementNode.cs" />
     <Compile Include="PathFindingCore\TileNodes\TileExceptionMetaData.cs" />
     <Compile Include="PathFindingCore\TileNodes\TileExceptionNode.cs" />
     <Compile Include="PathFindingCore\TileNodes\TileNodeObject.cs" />
     <Compile Include="PathFindingCore\Utilities.cs" />
-    <Compile Include="PathFindingCore\WaterLogic\WaterLogic.cs" />
+    <Compile Include="TaskCore\WaterLogic\WaterLogic.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
